Implement basic symbolic execution engine for quivela
This is a first implemention of symbolic-execution-based verification for
quivela programs. The symbolic evaluator produces a list of possible execution
paths for each method verification on both the left-hand and right-hand side
program. Each pair of paths then generates verification conditions expressing
method equivalence.

The conditions are discharged directly when trivial, or using z3 if they are
not. If z3 fails to verify them, Dafny is invoked as a fallback.