Prove that EtM provides authenticated encryption
This commit adds an example proof that Encrypt-then-Mac results in
authenticated encryption under the assumption that the underlying
encryption and macing primitives satisfy their respective security
properties.

The proof is embedded in a Haskell file, since quivela2 provides
verification as a deeply embedded DSL inside Haskell.

Note that since this showcases the compile-time verification
support, this adds about 10 seconds to the compilation time
of the full tool.