Preliminary support for calling methods on typed symbolic values
This sketches support for method calls where the receiver object is
symbolic. This only works if the type of the object is known to
have a type for which we have an effect specification of the method
being called.

In case the receiver object is a symbolic map lookup, we generate
two paths to explore (one path where the lookup succeeds and one
where it doesn't.