Check value against specified type
In the presence of object types that are specified in terms of the
effects of their methods, we can't the most general type of a value
anymore, so we need to instead compare a value against a desired type
directly. This commit implements functions checking this, though
they are not used anywhere at the moment.