Emit better representation for integers in Z3 backend
The verifier now emits terms of the form (VInt x) for symbolic variables that
have a Quivela type of integer to z3. This allows z3 to use its own builtin
arithmetic theories when reasoning about integer operations instead so that we
don't need to axiomatize arithmetic operations on quivela values ourselves.