Initialize new variables to 0.
This is needed since the FIFO proof sometimes assigns to a nested map index like
d[a][c] without the d[a] already existing in the map. To get uniform behavior,
we initialize both map entries and new variables to 0 when assigning to a
variable / map location that doesn't exist in the context yet.