Basic support for method calls on symbolic typed objects
This commit adds basic support for calling methods on objects that
are symbolic, but have a type for which we have a type declaration.
Currently this does not support universal invariants on symbolic
objects yet, since they require modifying the context and it's
tricky to figure out how to propagate that the invariants only
hold on the modified contexts.