Treat return values of new() as symbolic
We now treat the addresses of new() as any symbolic variables that may be
coerced into any other type of value (except that they don't collide and are
not errors). Since in the underlying semantics, new returns an arbitrary
bitstring, we could use the returned bitstring as a non-reference as well.

This also has the effect that we can use new(){} for random number generation.