Use SMTLib abstract syntax
Currently we have an "Emitter" that just dumps strings to Z3 as
it walks over the language elements.  This is ugly and hard to change.
Fortunately there are perfectly good SMT AST libraries.  This change
uses one of those libraries and translates verificaiton conditions to
SMTLIB2 abstract syntax and then to strings for Z3 to digest.

Note that this change causes the envelope encryption example to
fail, as we've removed a terrible hack in the code to add an
assertion.  This change is a prerequisite to a forthcoming change
that will give us the ability to write assertions in the concrete
Quivela syntax, and thereby allow us to fix the envelope example.
Otherwise all tests pass.

cr https://code.amazon.com/reviews/CR-4498481