Unify code for sets and maps
Sets and maps, even though nearly identical in behavior,
were handled somewhat differently in the code.  It seems
to me the set code was just buggy, but we don't have many
tests of the behavior.  We'll add some tests in a future
review.

cr https://code.amazon.com/reviews/CR-4649430