Use OSet to maintain a sequence of unique Prop instances instead of nub
Instead of repeatedly evaluating "nub" on lists, sequences of Props are
now managed using the Data.Set.OrderedSet.OSet (an
insertion-order-preserving set). This maintains the invariant that there
are no duplicate instances of Prop in the container and also retains the
order in which the Prop instances were inserted into it.

https://sim.amazon.com/issues/ARG-1047

cr https://code.amazon.com/reviews/CR-5177991